requirements.txt
numpy
matplotlib
streamlit
keras==2.10
tensorflow==2.10
opencv-python
Pillow
tensorflow_hub

